refactor: altera estrutura de client
