+++ lib/drasil.rb
@@ -12,6 +12,7 @@ require 'drasil/errors/resource_not_found_error'
 
 # Configuration and parsing
 require 'drasil/config'
+require 'drasil/config_resolver'
 require 'drasil/parser'
 require 'drasil/url_matcher'
 require 'drasil/url_pattern'
@@ -21,8 +22,7 @@ require 'drasil/response/raise_error'
 require 'drasil/json_parser'
 
 # Client-based architecture
-require 'drasil/client/configuration'
-require 'drasil/resource_registry'
+require 'drasil/client/context'
 require 'drasil/client'
 
 # Rails integration
